Another version of UNREPL blob
On this version, a big change: now, instead of printing the `pr-str`
representation of a plain old Java object, we use `as-str` in a way that
will disable elisions and try to parse it with edn/read-string. If it is
possible to parse it, it'll be the repr of the object.