Fixed SCI running code for vscode
Now, before you ask why I'm returning `nil`: the reasons is that VSCode
API is PURE GARBAGE. Really. Sometimes, some code will return a promise
that WILL NEVER RESOLVE, so other APIs that depend on the promise being
resolved WILL FAIL. Now, on this case, the `sci/read-string is
synchronous, so we don't need to return promises. But, I want to do it
in the future, so that reloading configs can't crash the editor if
something wrong happens - and, in this case, I'll have to REMEMBER that
VSCode programmers can't differentiate between doing IO and returning a
value, or even can't tie things correcly so that when you return a
promise that indicates that an IO is happening, you REMEMBER TO RESOLVE
IT when the IO completes!